0-print_list.c up1

+++ singly_linked_lists/0-print_list.c
@@ -10,7 +10,7 @@
 size_t print_list(const list_t *h)
 {
 
-	size_t node = 0;
+	int node = 0;
 
 	while (h != NULL)
 	{
